Mixed grid with flex to perfect the sidebar
I used grid to make the auto flow row then created smaller sections
within the sidebar. These sections were arranged in rows. Within the
sections, I created new divs to house the svg codes and text for each
tab in the sidebar.

The divs have a display of flex and an align items property of center to
make the text attached to the icon on the same line.